-updated classdiagramm: expense, timeslice -objectfactory implemented -classes optimised: expense, timeslice

+++ PlexByte.App.MoCap.Interactions/PlexByte.MoCap.Interactions/IAccount.cs
@@ -33,7 +33,7 @@ public interface IAccount
 
     void DeleteTimeslice(ITimeslice pTimeslice);
 
-    int UserExpense(IUser pUser);
+    decimal UserExpense(IUser pUser);
 
     int UserTimeslice(IUser pUser);
 }
